(back): create highlights subroute and cleanup code
