+++ backend/Dockerfile
@@ -19,11 +19,8 @@ RUN pip install --no-cache-dir -r requirements.txt
 # Copy the entire backend folder into the container
 COPY . /app
 
-# Change the working directory to /app/src for the app code
-WORKDIR /app/src
-
 # Expose the port on which the app will run
 EXPOSE 8000
 
 # Command to run the FastAPI server using Uvicorn
-CMD ["uvicorn", "server:app", "--host", "0.0.0.0", "--port", "8000", "--reload", "--reload-dir", "/app/src"]
+CMD ["uvicorn", "src.server:app", "--host", "0.0.0.0", "--port", "8000"]
